secured routes

+++ routes/order.js
@@ -25,9 +25,9 @@ orderRouter.post('/', function(req, res) {
 
 
 orderRouter.get('/', function (req, res) {
-    const limit = req.query.limit ? parseInt(req.query.limit) : undefined;
-    const offset = req.query.offset ? parseInt(req.query.offset) : undefined;
     if(UserController.verifyToken(req.headers['x-access-token']) === true){
+        const limit = req.query.limit ? parseInt(req.query.limit) : undefined;
+        const offset = req.query.offset ? parseInt(req.query.offset) : undefined;
         OrderController.getAll(req.query.name, limit, offset)
             .then((orders) => {
                 res.json(orders);
@@ -69,14 +69,18 @@ orderRouter.put('/:id/addmenu/:menuId', function (req, res) {
 });
 
 orderRouter.delete('/:id', function (req, res) {
-    const orderId = parseInt(req.params.id);
-    OrderController.delete(orderId)
-        .then((order) => {
-            res.status(200).json(order.deletedAt);
-        })
-        .catch((err) => {
-            res.status(500).end();
-        });
+    if(UserController.verifyToken(req.headers['x-access-token']) === true){
+        const orderId = parseInt(req.params.id);
+        OrderController.delete(orderId)
+            .then((order) => {
+                res.status(200).json(order.deletedAt);
+            })
+            .catch((err) => {
+                res.status(500).end();
+            });
+    }else{
+        res.status(403).end();
+    }
 });
 
 module.exports = orderRouter;
